Refactor iterator into modules.
