+++ src/core/decoder/mod.rs
@@ -1,2 +1,3 @@
 pub mod decoder;
+pub mod gif;
 pub mod image;
